+++ eth/transaction.go
@@ -2,14 +2,9 @@ package eth
 
 import (
 	"encoding/json"
-
-	"github.com/pkg/errors"
-
-	"github.com/INFURA/go-ethlibs/rlp"
 )
 
 type Condition json.RawMessage
-type AccessList []AccessListEntry
 
 var (
 	TransactionTypeLegacy     = int64(0x0)
@@ -46,11 +41,6 @@ type Transaction struct {
 
 	// Keep the source so we can recreate its expected representation
 	source string
-}
-
-type AccessListEntry struct {
-	Address     Address  `json:"address"`
-	StorageKeys []Data32 `json:"storageKeys"`
 }
 
 type NewPendingTxBodyNotificationParams struct {
@@ -165,32 +155,4 @@ func (t *Transaction) MarshalJSON() ([]byte, error) {
 	type unknown Transaction
 	u := unknown(*t)
 	return json.Marshal(&u)
-}
-
-func NewAccessListFromRLP(v rlp.Value) (AccessList, error) {
-	accessList := make(AccessList, len(v.List))
-	for j, accessRLP := range v.List {
-		l := len(accessRLP.List)
-		if l == 0 || l > 2 {
-			return nil, errors.Errorf("invalid access list entry %d", j)
-		}
-		address, err := NewAddress(accessRLP.List[0].String)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid access list entry address %d", j)
-		}
-		accessList[j].Address = *address
-		if l == 2 {
-			// 2nd item is the storage keys
-			accessList[j].StorageKeys = make([]Data32, len(accessRLP.List[1].List))
-			for k, key := range accessRLP.List[1].List {
-				d, err := NewData32(key.String)
-				if err != nil {
-					return nil, errors.Wrapf(err, "invalid access list entry %d storage key %d", j, k)
-				}
-				accessList[j].StorageKeys[k] = *d
-			}
-		}
-	}
-
-	return accessList, nil
 }
